% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_tests.R
\name{apply_test.test_many_many}
\alias{apply_test.test_many_many}
\title{Tests One to One}
\usage{
\method{apply_test}{test_many_many}(df_left, df_right, setup)
}
\arguments{
\item{df_left}{left dataset}

\item{df_right}{right dataset}

\item{setup}{from the merge setup class}
}
\description{
Tests One to One
}
