% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_tests.R
\name{apply_test.test_orphan_rec}
\alias{apply_test.test_orphan_rec}
\title{Tests One to Many}
\usage{
\method{apply_test}{test_orphan_rec}(primary_df, related_df, setup)
}
\arguments{
\item{setup}{from the merge setup class}

\item{df_left}{left dataset}

\item{df_right}{right dataset}
}
\description{
Tests One to Many
}
