% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{apply_tests}
\alias{apply_tests}
\title{Apply tests}
\usage{
apply_tests(df, setup)
}
\arguments{
\item{df}{the dataframe to apply the test to}

\item{setup}{the test setup}
}
\value{
A dataframe representing the values from a single test
}
\description{
Applied the tests
}
\examples{
df <- data.frame(x = 1:10, y = c(rep("X", 4), rep("Y", 4), rep("Z", 2)))
setup1 <- list(test_name = "test_range", df_name = "Test_df", col_name = "x",
               int = TRUE, lower_inclu = TRUE, upper_inclu = TRUE, lower = 1,
               upper = 10, na = FALSE)

setup2 <- list(test_name = "test_values", df_name = "Test_df", col_name = "y",
               values = c("X", "Y"), na = FALSE)
test1 <- apply_tests(df, setup1)
test2 <- apply_tests(df, setup2)
}
