% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{assign_numeric_class}
\alias{assign_numeric_class}
\title{Assign the correct numeric test}
\usage{
assign_numeric_class(df, cls)
}
\arguments{
\item{df}{the dataframe}

\item{cls}{the setup class for testing}
}
\description{
Assigns the correct numeric test depending on the parameters
}
