% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_test_classes.R
\name{class_test_bool}
\alias{class_test_bool}
\title{Class to test boolean}
\usage{
class_test_bool(setup)
}
\arguments{
\item{setup}{the general test setup}
}
\description{
Class to test boolean
}
\details{
the setup class should have three parameters: df_name, col_name, and na.
df_name is a string and represents the name of the dataframe. 
col_name is a string and represents the name of the column to be tested.
na is a boolean value, if NA values are allowed in the column to be tested. If NA = FALSE and
there are NA values, an error will occur.
}
\examples{
setup <- list(df_name = "data name", col_name = "col name", na = TRUE)
cls <- class_test_bool(setup) 
}
