% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_test_classes.R
\name{class_test_na}
\alias{class_test_na}
\title{Class to test NA}
\usage{
class_test_na(setup)
}
\arguments{
\item{setup}{the general test setup}
}
\description{
Class to test NA
}
\details{
the setup class should have two parameters: df_name, col_name, values, and na.
df_name is a string and represents the name of the dataframe. 
col_name is a string and represents the name of the column to be tested.
}
\examples{
setup <- list(df_name = "data name", col_name = "col name")
cls <- class_test_na(setup)
}
