% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_test_classes.R
\name{class_test_values}
\alias{class_test_values}
\title{Class to test values}
\usage{
class_test_values(cls)
}
\arguments{
\item{cls}{the general test setup cls}
}
\description{
Class to test values
}
\details{
if cls$na is TRUE then it gets added to the list of acceptable values
}
