% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{max_na}
\alias{max_na}
\title{Max with NA}
\usage{
max_na(col, na)
}
\arguments{
\item{col}{the column to test}

\item{na}{if the column should include NA values or not, takes the values TRUE or FALSE}
}
\description{
Returns the maximum with NA's removed if na = TRUE
}
