% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_range.R
\name{min_na}
\alias{min_na}
\title{Min with NA}
\usage{
min_na(col, na)
}
\arguments{
\item{col}{the column to test}

\item{na}{if the column should include NA values or not, takes the values TRUE or FALSE}
}
\description{
Returns the minimum with NA's removed if na = TRUE
}
