% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_merges.R
\name{setup_test_merge}
\alias{setup_test_merge}
\title{Setup Test merges}
\usage{
setup_test_merge(df_left_name, df_right_name, left_on, right_on,
  left_na = FALSE, right_na = FALSE)
}
\arguments{
\item{df_left_name}{the name of left dataset}

\item{df_right_name}{right dataset}

\item{left_on}{the column or vector of columns to merge on}

\item{right_on}{the column or vector of columns to merge on}

\item{left_na}{from the left dataset are okay, default = FALSE}

\item{right_na}{from the right dataset are okay, default = FALSE}
}
\description{
Setup Test merges
}
