% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_test_classes.R
\name{setup_test_na}
\alias{setup_test_na}
\title{Setup Test Na}
\usage{
setup_test_na(df_name, col_name)
}
\arguments{
\item{df_name}{is a string and represents the name of the dataframe.}

\item{col_name}{is a string and represents the name of the column to be tested.}
}
\description{
Setup Test Na
}
