% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_merges.R
\name{setup_test_orphan_rec}
\alias{setup_test_orphan_rec}
\title{Setup for tests for orphaned records}
\usage{
setup_test_orphan_rec(primary_df, related_df, primary_key, foreign_key)
}
\arguments{
\item{primary_df}{the name of the primary dataframe}

\item{related_df}{the name of the related dataframe}

\item{primary_key}{the column or vector of columns to merge on}

\item{foreign_key}{the column or vector of columns to merge on}
}
\description{
Setup for tests for orphaned records
}
