% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_unique.R
\name{setup_test_unique}
\alias{setup_test_unique}
\title{Setup Test Unique}
\usage{
setup_test_unique(df_name, col_name, na)
}
\arguments{
\item{df_name}{is a string and represents the name of the dataframe.}

\item{col_name}{is a string and represents the name of the column to be tested.}

\item{na}{is a boolean value, if NA values are allowed in the column to be tested. If NA = FALSE and
there are NA values, an error will occur.}
}
\description{
Setup Test Unique
}
\examples{
df <- data.frame(x = 1:10, y = c(rep("X", 4), rep("Y", 4), rep("Z", 2)))
setup <- setup_test_unique(df_name = "df", col_name = "x", na = FALSE)
}
