% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_test_classes.R
\name{setup_test_values}
\alias{setup_test_values}
\title{Setup Test Values}
\usage{
setup_test_values(df_name, col_name, values, na)
}
\arguments{
\item{df_name}{is a string and represents the name of the dataframe.}

\item{col_name}{is a string and represents the name of the column to be tested.}

\item{values}{is a list of values that expected in the column to be tested.}

\item{na}{is a boolean value, if NA values are allowed in the column to be tested. If NA = FALSE and
there are NA values, an error will occur.}
}
\description{
Setup Test Values
}
