% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test.exclu_lower}
\alias{test.exclu_lower}
\title{Test Exclu lower}
\usage{
\method{test}{exclu_lower}(setup, df, ...)
}
\arguments{
\item{setup}{the setup class}

\item{df}{the dataframe}

\item{...}{other arguments passed to the function}
}
\description{
Tests that column is greater than the lower bound
}
