% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test.exclu_lower_inclu_upper}
\alias{test.exclu_lower_inclu_upper}
\title{Test exclu lower and inclu upper}
\usage{
\method{test}{exclu_lower_inclu_upper}(setup, df, ...)
}
\arguments{
\item{setup}{the setup class}

\item{df}{the dataframe}
}
\description{
Tests that the column is greater than the lower bound AND less or equal to than the upper bound
}
