% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test.exclu_upper}
\alias{test.exclu_upper}
\title{Test Exclu upper}
\usage{
\method{test}{exclu_upper}(setup, df, ...)
}
\arguments{
\item{setup}{the setup class}

\item{df}{the dataframe}

\item{...}{other arguments passed to the function}
}
\description{
Tests that column is less than the upper bound
}
