% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test.inclu_lower}
\alias{test.inclu_lower}
\title{Test inclu lower}
\usage{
\method{test}{inclu_lower}(setup, df, ...)
}
\arguments{
\item{setup}{the setup class}

\item{df}{the dataframe}

\item{...}{other arguments passed to the function}
}
\description{
Tests that column is greater than or equal to the lower bound
}
