% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test.inclu_lower_exclu_upper}
\alias{test.inclu_lower_exclu_upper}
\title{Test inclu lower and exclu upper}
\usage{
\method{test}{inclu_lower_exclu_upper}(setup, df, ...)
}
\arguments{
\item{setup}{the setup class}

\item{df}{the dataframe}

\item{...}{other arguments passed to the function}
}
\description{
Tests that the column is greater than or equal to the lower bound AND less than the upper bound
}
