% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test.inclu_upper}
\alias{test.inclu_upper}
\title{Test Exclu upper}
\usage{
\method{test}{inclu_upper}(setup, df, ...)
}
\arguments{
\item{setup}{the setup class}

\item{df}{the dataframe}
}
\description{
Tests that column is less than or equal to the upper bound
}
