% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test.na}
\alias{test.na}
\title{Test NA values}
\usage{
\method{test}{na}(setup, df, ...)
}
\arguments{
\item{setup}{the setup class}

\item{df}{the dataframe}
}
\description{
Tests if the vector/column contains any NA values
}
\examples{
\dontrun{
df <- data.frame(x = 1:4, y = c(NA, 6:8))
setup <- setup_test_na("df_name", "x") 
test <- test(setup, df)
## test$test_result returns TRUE

setup <- setup_test_na("df_name", "y") 
test <- test(setup, df)
## test$test_result returns FALSE
}
}
