% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test.orphan_rec}
\alias{test.orphan_rec}
\title{Tests for orphaned records}
\usage{
\method{test}{orphan_rec}(setup, primary_df, related_df, ...)
}
\arguments{
\item{setup}{the setup class}

\item{primary_df}{the primary dataframe}

\item{related_df}{the related dataframe}
}
\description{
Tests that there is a one to many relationship between the left and right dataset
}
\examples{
primary_df <- data.frame(id1 = c(1,2,2,3), id2 = c(1,1,2,1), y = c(1,1,2,5))
related_df <- data.frame(id1 = c(1,2,3), id2 = c(1,1,1), x = c(1,2,3))

setup <- setup_test_orphan_rec(primary_df = "df1", related_df = "df2", 
                               primary_key = c("id1", "id2"), foreign_key = c("id1", "id2"))
test <- test(setup, primary_df, related_df)
## test$test_result returns TRUE
}
