% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test.range}
\alias{test.range}
\title{Assign the correct numeric test}
\usage{
\method{test}{range}(setup, df, ...)
}
\arguments{
\item{setup}{the setup class}

\item{df}{the dataframe}
}
\description{
Assigns the correct numeric test depending on the parameters
}
