% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test.unique}
\alias{test.unique}
\title{Test uniqueness}
\usage{
\method{test}{unique}(setup, df, ...)
}
\arguments{
\item{setup}{the setup class}

\item{df}{the dataframe}

\item{...}{other arguments passed to the function}
}
\description{
Tests if the vector/column is unique
}
\examples{
\dontrun{
df <- data.frame(x = 1:4, y = c(1,1,2:3))
setup <- setup_test_unique("df_name", "x") 

test <- test(setup, df)
## test$test_result returns TRUE

setup <- setup_test_unique("df_name", "y") 
test <- test(setup, df)
## test$test_result returns FALSE
}
}
