% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test.values}
\alias{test.values}
\title{Test expected values}
\usage{
\method{test}{values}(setup, df, ...)
}
\arguments{
\item{setup}{the setup class}

\item{df}{the dataframe}
}
\value{
class with test_result, test_message, test_description
}
\description{
Tests if the vector/column contains values other than expected
}
\examples{
\dontrun{
df <- data.frame(x = 1:4, y = 5:8)
setup <- setup_test_values("df_name", "x", values = 1:4, na = FALSE)

test <- test(setup, df)
## test$test_result returns TRUE

setup <- setup_test_values("df_name", "y", values = 1:4, na = FALSE)
test <- test(setup, df)
## test$test_result returns FALSE
}
}
