% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bool.R
\name{test_all_false}
\alias{test_all_false}
\title{Tests all rows equal FALSE}
\usage{
test_all_false(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
vector
}
\description{
Tests all rows equal FALSE
}
\examples{
vec <- c(T, F, T, F)
test_all_false(vec)[2]
# FAILS
df <- data.frame(col1 = c(F, F, F, F))
test_all_false(df$col1)[2]
# PASSES
}
