% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bool.R
\name{test_all_true_test}
\alias{test_all_true_test}
\title{Tests all values equal true}
\usage{
test_all_true_test(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
boolean
}
\description{
Tests all values equal true
}
\examples{
vec <- c(T, F, T, F)
test_all_true(vec)
# FAILS
df <- data.frame(col1 = c(T, T, T, T))
test_all_true(df$col1)
# PASSES
}
