% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bool.R
\name{test_any_true}
\alias{test_any_true}
\title{Tests any rows equal TRUE}
\usage{
test_any_true(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
vector
}
\description{
Tests any rows equal TRUE
}
\examples{
vec <- c(T, F, T, F)
test_any_true(vec)[2]
# PASSES
df <- data.frame(col1 = c(T, T, T, T))
test_any_true(df$col1)[2]
# PASSES
}
