% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bool.R
\name{test_any_true_test}
\alias{test_any_true_test}
\title{Tests any values equal TRUE}
\usage{
test_any_true_test(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
boolean
}
\description{
Tests any values equal TRUE
}
\examples{
vec <- c(T, F, T, F)
test_any_true(vec)
## Returns TRUE
df <- data.frame(col1 = c(T, T, T, T))
test_any_true(df$col1)
## Returns TRUE
}
