% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_conditional}
\alias{test_conditional}
\title{Test conditional}
\usage{
test_conditional(setup, t)
}
\arguments{
\item{setup}{the setup class}

\item{t}{a boolean value of TRUE or FALSE, if t == TRUE then return the passing class, else return the failing class}
}
\description{
Applies the attribute of a passing or failing test
}
