% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{test_double_range}
\alias{test_double_range}
\title{Test double range}
\usage{
test_double_range(df, col_name, upper_inclu, lower_inclu, upper, lower, na)
}
\arguments{
\item{df}{the dataframe}

\item{col_name}{the column name}

\item{upper_inclu}{indicates if the upper bound should be inclusive or not, takes on the values NULL, TRUE or FALSE}

\item{lower_inclu}{indicates if the lower bound should be inclusive or not, takes on the values NULL, TRUE or FALSE}

\item{upper}{the upper value}

\item{lower}{the lower value}

\item{na}{if the column should include NA values or not, takes the values TRUE or FALSE}
}
\description{
Tests the double range based on the parameter configurations
}
