% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{test_exclu_lower}
\alias{test_exclu_lower}
\title{Test Exclu lower}
\usage{
test_exclu_lower(df, class)
}
\arguments{
\item{lower}{the lower bound}
}
\description{
Tests that column is greater than the lower bound
}
