% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_setup.R
\name{test_exclu_lower_exclu_upper}
\alias{test_exclu_lower_exclu_upper}
\title{Test exclu lower and exclu upper}
\usage{
test_exclu_lower_exclu_upper(df, setup)
}
\arguments{
\item{df}{the dataframe}

\item{setup}{the setup class for testing}
}
\description{
Tests that the column is greater than the lower bound AND less than the upper bound
}
