% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{test_exclu_lower_inclu_upper}
\alias{test_exclu_lower_inclu_upper}
\title{Test exclu lower and inclu upper}
\usage{
test_exclu_lower_inclu_upper(df, class)
}
\description{
Tests that the column is greater thanto the lower bound AND less or equal to than the upper bound
}
