% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{test_exclu_lower_inclu_upper_range}
\alias{test_exclu_lower_inclu_upper_range}
\title{Test exclusive lower bound, inclusive upper bound}
\usage{
test_exclu_lower_inclu_upper_range(vec, lower, upper)
}
\arguments{
\item{lower}{the lower bound}

\item{upper}{the upper bound}
}
\value{
vector
}
\description{
Tests if all the values are less than the upper and greater than the lower bound or equal to upper/lower bound
}
\examples{
vec <- c(1,2,3)

test_exclu_lower_inclu_upper_range(vec, 0, 3)
## Returns PASS

test_exclu_lower_inclu_upper_range(vec, 1, 3)
## Returns ERROR

test_exclu_lower_inclu_upper_range(vec, 0, 2)
## Returns ERROR
}
\seealso{
Other numeric column tests: \code{\link{test_exclu_value_range}},
  \code{\link{test_greater_than_or_equal_value}},
  \code{\link{test_greater_than_value}},
  \code{\link{test_inclu_value_range}},
  \code{\link{test_less_than_or_equal_value}},
  \code{\link{test_less_than_value}}
}
\concept{numeric column tests}
