% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{test_exclu_upper}
\alias{test_exclu_upper}
\title{Test Exclu upper}
\usage{
test_exclu_upper(df, class)
}
\arguments{
\item{col}{the column to test}

\item{upper}{the upper bound}

\item{na}{if the column should include NA values or not, takes the values TRUE or FALSE}
}
\description{
Tests that column is less than the upper bound
}
