% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{test_exclu_value_range}
\alias{test_exclu_value_range}
\title{Test exclusive bounds range}
\usage{
test_exclu_value_range(vec, lower, upper)
}
\arguments{
\item{vec}{the vector or column to test}

\item{lower}{the lower bound}
}
\value{
vector
}
\description{
Tests if all the values are less than the upper and greater than the lower bound
}
\examples{
vec <- c(1,2,3,4,5)

test_exclu_value_range(vec, 2, 4)
# Returns PASS

test_exclu_value_range(vec, 1, 4)
## Returns ERROR

test_exclu_value_range(vec, 2, 5)
## Returns ERROR
}
\seealso{
Other numeric column tests: \code{\link{test_greater_than_or_equal_value}},
  \code{\link{test_greater_than_value}},
  \code{\link{test_less_than_or_equal_value}},
  \code{\link{test_less_than_value}}
}
\concept{numeric column tests}
