% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{test_exclu_value_range_test}
\alias{test_exclu_value_range_test}
\title{Test exclusive bounds range}
\usage{
test_exclu_value_range_test(vec, lower, upper)
}
\arguments{
\item{lower}{the lower bound}

\item{upper}{the upper bound}
}
\value{
boolean
}
\description{
Tests if all the values in a column between an upper and lower bound
}
\examples{
vec <- c(1,2,3)

test_exclu_exclu_range_test(vec, 0, 4)
## Returns TRUE

test_exclu_exclu_range_test(vec, 0, 3)
## Returns FALSE

test_exclu_exclu_range_test(vec, 1, 3)
## Returns FALSE
}
