% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{test_greater_than_or_equal_value}
\alias{test_greater_than_or_equal_value}
\title{Test greater than lower bound}
\usage{
test_greater_than_or_equal_value(vec, lower)
}
\arguments{
\item{vec}{the vector or column to test}

\item{lower}{the lower bound}
}
\value{
vector
}
\description{
Tests if all the values in a column are greater than an lower bound
}
\examples{
vec <- c(1,2,3)

test_greater_than_value(vec, 0)[2]
# Returns PASS

test_greater_than_value(vec, 1)[2]
## Returns PASS

test_greater_than_value(vec, 2)[2]
## Returns ERROR
}
\seealso{
Other numeric column tests: \code{\link{test_exclu_value_range}},
  \code{\link{test_greater_than_value}},
  \code{\link{test_inclu_value_range}},
  \code{\link{test_less_than_or_equal_value}},
  \code{\link{test_less_than_value}}
}
\concept{numeric column tests}
