% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{test_greater_than_value_test}
\alias{test_greater_than_value_test}
\title{Test greater than lower bound}
\usage{
test_greater_than_value_test(vec, lower)
}
\arguments{
\item{vec}{the vector or column to test}

\item{lower}{the lower bound}
}
\value{
boolean
}
\description{
Tests if all the values in a column are greater than a lower bound
}
\examples{
vec <- c(1,2,3)

test_greater_than_value_test(vec, 0)
# Returns TRUE

test_greater_than_value_test(vec, 1)
## Returns FALSE
}
