% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{test_inclu_lower_exclu_upper_range_test}
\alias{test_inclu_lower_exclu_upper_range_test}
\title{Test inclusive lower bound, exclusive upper bound}
\usage{
test_inclu_lower_exclu_upper_range_test(vec, lower, upper)
}
\arguments{
\item{vec}{the vector or column to test}

\item{lower}{the lower bound}

\item{upper}{the upper bound}
}
\value{
boolean
}
\description{
Tests if all the values in a column are less than or equal to the lower bound and greater than the upper bound
}
\examples{
vec <- c(1,2,3)

test_inclu_lower_exclu_upper_range_test (vec, 1, 4)
## Returns TRUE

test_inclu_lower_exclu_upper_range_test(vec, 1, 3)
## Returns FALSE

test_inclu_lower_exclu_upper_range_test(vec, 2, 4)
## Returns FALSE
}
