% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{test_inclu_upper}
\alias{test_inclu_upper}
\title{Test Exclu upper}
\usage{
test_inclu_upper(df, cls)
}
\arguments{
\item{df}{the dataframe}

\item{cls}{the setup class for testing}
}
\description{
Tests that column is less than or equal to the upper bound
}
