% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{test_less_than_or_equal_value}
\alias{test_less_than_or_equal_value}
\title{Test less than or equal to upper bound}
\usage{
test_less_than_or_equal_value(vec, upper)
}
\arguments{
\item{vec}{the vector or column to test}

\item{upper}{the upper bound}
}
\value{
vector
}
\description{
Tests if all the values in a column are less than or equal to an upper bound
}
\examples{
vec <- c(1,2,3)

test_less_than_or_equal_value_test(vec, 4)[2]
## Returns PASS

test_less_than_or_equal_value_test(vec, 3)[2]
## Returns PASS

test_less_than_or_equal_value_test(vec, 2)[2]
## Returns ERROR
}
\seealso{
Other numeric column tests: \code{\link{test_exclu_lower_inclu_upper_range}},
  \code{\link{test_exclu_value_range}},
  \code{\link{test_greater_than_or_equal_value}},
  \code{\link{test_greater_than_value}},
  \code{\link{test_inclu_lower_exclu_upper_range}},
  \code{\link{test_inclu_value_range}},
  \code{\link{test_less_than_value}}
}
\concept{numeric column tests}
