% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{test_less_than_or_equal_value_test}
\alias{test_less_than_or_equal_value_test}
\title{Test less than or equal to upper bound}
\usage{
test_less_than_or_equal_value_test(vec, upper)
}
\description{
Test less than or equal to upper bound
}
