% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{test_less_than_value_test}
\alias{test_less_than_value_test}
\title{Test less than upper bound}
\usage{
test_less_than_value_test(vec, upper)
}
\arguments{
\item{vec}{the vector or column to test}

\item{upper}{the upper bound}
}
\value{
boolean
}
\description{
Tests if all the values in a column are less than an upper bound
}
\examples{
\dontrun{
vec <- c(1,2,3)

test_less_than_value_test(vec, 4)
# Returns TRUE

test_less_than_value_test(vec, 3)
## Returns FALSE
}
}
