% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_merges.R
\name{test_many_many}
\alias{test_many_many}
\title{Test one to one}
\usage{
test_many_many(df_left, df_right, setup)
}
\arguments{
\item{df_left}{left dataset}

\item{df_right}{right dataset}

\item{setup}{from the merge setup class}
}
\description{
Tests that there is a many to many relationship between the left and right dataset
}
