% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{test_merge}
\alias{test_merge}
\title{Create row for test merge}
\usage{
test_merge(test)
}
\arguments{
\item{test}{the test class generated by one of the test functions}
}
\description{
Create row for test merge
}
\examples{
test <- test_one_one(df_left, df_right, setup)
test_merge(test)
}
