% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_col.R
\name{test_na}
\alias{test_na}
\title{Test NA values}
\usage{
test_na(df, col_name)
}
\arguments{
\item{df}{the dataframe}

\item{col_name}{the column name}
}
\description{
Tests if the vector/column contains any NA values
}
\examples{
\dontrun{
df <- data.frame(x = 1:4, y = c(NA, 6:8))

test_na(df, "x")
## test$test_result returns TRUE

test_na(df, "y")
## test$test_result returns FALSE
}
}
