% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_col.R
\name{test_null_values}
\alias{test_null_values}
\title{Test NA values}
\usage{
test_null_values(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
vector
}
\description{
Tests if the vector/column contains any NA values
}
\examples{
vec1 <- c(1,2,3)
vec2 <- c(1,NA,2)

test_null_values(vec1)[2]
## Returns PASS

test_null_values(vec2)[2]
## Returns ERROR 
}
\seealso{
Other general column tests: \code{\link{test_unique}},
  \code{\link{test_values}}
}
\concept{general column tests}
