% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_col.R
\name{test_null_values_test}
\alias{test_null_values_test}
\title{Test NA values}
\usage{
test_null_values_test(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
boolean
}
\description{
Tests if the vector/column contains any NA values
}
\examples{
\dontrun{
vec1 <- c(1,2,3)
vec2 <- c(1,NA,2)

test_null_values_test(vec1)
# Returns TRUE

test_null_values_test(vec2)
## Returns FALSE
}
}
