% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_merges.R
\name{test_one_many}
\alias{test_one_many}
\title{Test one to many}
\usage{
test_one_many(df_left, df_right, setup)
}
\arguments{
\item{df_left}{left dataset}

\item{df_right}{right dataset}

\item{setup}{from the merge setup class}
}
\description{
Tests that there is a one to many relationship between the left and right dataset
}
