% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_merges.R
\name{test_orphan_rec}
\alias{test_orphan_rec}
\title{Tests for orphaned records}
\usage{
test_orphan_rec(primary_df, related_df, setup)
}
\arguments{
\item{primary_df}{the primary dataframe}

\item{related_df}{the related dataframe}

\item{setup}{from the merge setup class}
}
\description{
Tests that there is a one to many relationship between the left and right dataset
}
