% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{test_param_logical_or_na}
\alias{test_param_logical_or_na}
\title{Test param is logical or NA}
\usage{
test_param_logical_or_na(setup, param)
}
\arguments{
\item{setup}{the class to check the parameters}

\item{param}{string, the parameter to check}
}
\description{
Test param is logical or NA
}
