% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{test_param_numeric}
\alias{test_param_numeric}
\title{Test param is numeric}
\usage{
test_param_numeric(setup, param)
}
\arguments{
\item{setup}{the class to check the parameters}

\item{param}{string, the parameter to check}
}
\description{
Test param is numeric
}
