% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_test_classes.R
\name{test_param_string}
\alias{test_param_string}
\title{Test param is string}
\usage{
test_param_string(cls, param)
}
\arguments{
\item{cls}{the class to check the parameters}

\item{param}{string, the parameter to check}
}
\description{
Test param is string
}
