% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{test_params_both_na_or_not}
\alias{test_params_both_na_or_not}
\title{Test params are both na or both not na}
\usage{
test_params_both_na_or_not(setup, param1, param2)
}
\arguments{
\item{setup}{the class to check the parameters}

\item{param1}{first param to test}

\item{param2}{second param to test}
}
\description{
Tests that param1 and param2 both equal na or both are not equal to na
}
