% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_test_classes.R
\name{test_params_both_not_null}
\alias{test_params_both_not_null}
\title{Test two params are both not null}
\usage{
test_params_both_not_null(cls, param1, param2)
}
\description{
Test two params are both not null
}
