% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_test_classes.R
\name{test_params_both_null_or_not}
\alias{test_params_both_null_or_not}
\title{Test params are both null or both not null}
\usage{
test_params_both_null_or_not(cls, param1, param2)
}
\arguments{
\item{param1}{first param to test}

\item{param2}{second param to test}
}
\description{
Tests that param1 and param2 both equal null or both are not equal to null
}
