% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_range.R
\name{test_range}
\alias{test_range}
\title{Assign the correct numeric test}
\usage{
test_range(df, cls)
}
\arguments{
\item{df}{the dataframe}

\item{cls}{the setup class for testing}
}
\description{
Assigns the correct numeric test depending on the parameters
}
