% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{test_row}
\alias{test_row}
\title{Create test row}
\usage{
test_row(test)
}
\arguments{
\item{test}{the test class generated by one of the test functions}
}
\description{
Create test row
}
\examples{
test <- test_na(df, cls)
test_row(test)
}
