% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_tests.R
\name{test_summary.column}
\alias{test_summary.column}
\title{Creates a summary of the test}
\usage{
\method{test_summary}{column}(setup, ...)
}
\arguments{
\item{setup}{the setup class}
}
\value{
dataframe
}
\description{
Creates a summary of the test
}
