% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_tests.R
\name{test_summary.merge}
\alias{test_summary.merge}
\title{Create row for test merge}
\usage{
\method{test_summary}{merge}(setup, ...)
}
\arguments{
\item{test}{the test class generated by one of the test functions}
}
\description{
Create row for test merge
}
