% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_tests.R
\name{test_summary}
\alias{test_summary}
\title{Get summary}
\usage{
test_summary(setup, ...)
}
\arguments{
\item{setup}{the setup class}
}
\value{
dataframe
}
\description{
Summarizes the attributes from the test
}
