% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_col.R
\name{test_unique}
\alias{test_unique}
\title{Test uniqueness}
\usage{
test_unique(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
vector
}
\description{
Tests if the vector/column is unique
}
\examples{
vec1 <- c(1,2,3)
vec2 <- c(1,1,2)

test_unique(vec1)[2]
## Returns PASS

test_unique(vec2)[2]
## Returns ERROR 
}
\seealso{
Other general column tests: \code{\link{test_null_values}},
  \code{\link{test_values}}
}
\concept{general column tests}
