% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_col.R
\name{test_unique_test}
\alias{test_unique_test}
\title{Test uniqueness}
\usage{
test_unique_test(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
boolean
}
\description{
Tests if the vector/column is unique
}
\examples{
\dontrun{
vec1 <- c(1,2,3)
vec2 <- c(1,1,2)

test_unique_test(vec1)
# Returns TRUE

test_unique_test(vec2)
## Returns FALSE
}
}
