% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_col.R
\name{test_values_test}
\alias{test_values_test}
\title{Test expected values}
\usage{
test_values_test(add_values)
}
\arguments{
\item{add_values}{a list additional values}
}
\value{
boolean
}
\description{
Tests if the vector/column contains values other than expected
}
\examples{
\dontrun{
test_values_test(c())
# Returns TRUE

test_values_test(c(4,5))
## Returns FALSE
}
}
