% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_tests.R
\name{wrong_rows}
\alias{wrong_rows}
\title{Get Problems}
\usage{
wrong_rows(test, ...)
}
\arguments{
\item{test}{the test class}

\item{...}{other arguments passed to the function}
}
\value{
dataframe
}
\description{
Get Problems
}
