% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_test_classes.R
\name{class_test_na}
\alias{class_test_na}
\title{Class to test NA}
\usage{
class_test_na(df_name, col_name)
}
\arguments{
\item{df_name}{the dataframe name}

\item{col_name}{the column name}
}
\description{
Class to test NA
}
