% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_test_classes.R
\name{class_test_range}
\alias{class_test_range}
\title{Class for range}
\usage{
class_test_range(df_name, col_name, int, upper_inclu, lower_inclu, upper,
  lower, na)
}
\description{
Class for range
}
