% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_test_classes.R
\name{class_test_unique}
\alias{class_test_unique}
\title{Class to test unique}
\usage{
class_test_unique(col_name)
}
\arguments{
\item{col_name}{the column name}
}
\description{
Class to test unique
}
