% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_test_classes.R
\name{class_test_unique}
\alias{class_test_unique}
\title{Class to test unique}
\usage{
class_test_unique(cls)
}
\arguments{
\item{cls}{the general test setup cls}
}
\description{
Class to test unique
}
\details{
if NA is set to TRUE, columns with multiple NAs will not be marked as non-unique, NA rows will be ignored
}
