% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_test_classes.R
\name{class_test_values}
\alias{class_test_values}
\title{Class for column type string}
\usage{
class_test_values(df_name, col_name, values, na)
}
\arguments{
\item{df_name}{the dataframe name}

\item{col_name}{the column name}

\item{values}{the expected values for the variable}

\item{na}{if the column should include NA values or not, takes the values TRUE or FALSE}
}
\description{
Class for column type string
}
\details{
if NA is TRUE then it gets added to the list of acceptable values
}
