% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_setup.R
\name{setup_test_na}
\alias{setup_test_na}
\title{Setup Test Na}
\usage{
setup_test_na(df_name, col_name)
}
\arguments{
\item{df_name}{is a string and represents the name of the dataframe.}

\item{col_name}{is a string and represents the name of the column to be tested.}
}
\description{
Setup Test Na
}
\examples{
df <- data.frame(x = 1:10, y = c(rep("X", 4), rep("Y", 4), rep("Z", 2)))
setup <- setup_test_na(df_name = "df", col_name = "y")
}
