% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_test_classes.R
\name{setup_test_range}
\alias{setup_test_range}
\title{Setup Test Range}
\usage{
setup_test_range(df_name, col_name, int, lower_inclu, upper_inclu, lower,
  upper, na)
}
\arguments{
\item{df_name}{is a string and represents the name of the dataframe.}

\item{col_name}{is a string and represents the name of the column to be tested.}

\item{int}{is a boolean value, if the values should be integers or not}

\item{lower_inclu}{takes on the values TRUE, FALSE, or NULL. If there's an lower bound then it should be set to TRUE or FALSE.
If the lower bound is inclusive, x >= lower, then lower_inclu = TRUE, else lower_inclu = FALSE.}

\item{upper_inclu}{takes on the values TRUE, FALSE, or NULL. If there's an upper bound then it should be set to TRUE or FALSE.
If the upper bound is inclusive, x <= upper, then upper_inclu = TRUE, else upper_inclu = FALSE.}

\item{lower}{can take a numeric value or NULL. If there's no lower bound, then lower should be null.}

\item{upper}{can take a numeric value or NULL. If there's no upper bound, then upper should be null.}

\item{na}{is a boolean value, if NA values are allowed in the column to be tested. If NA = FALSE and
there are NA values, an error will occur.}
}
\description{
Setup Test Range
}
