% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_setup.R
\name{setup_test_values}
\alias{setup_test_values}
\title{Setup Test Values}
\usage{
setup_test_values(df_name, col_name, values, ptr = NULL)
}
\arguments{
\item{df_name}{is a string and represents the name of the dataframe.}

\item{col_name}{is a string and represents the name of the column to be tested.}

\item{values}{is a list of values that expected in the column to be tested.}

\item{ptr}{a string or vector of strings correspond to the unique identifier or unique identifiers of the dataset}
}
\description{
Setup Test Values
}
\examples{
df <- data.frame(x = 1:10, y = c(rep("X", 4), rep("Y", 4), rep("Z", 2)))
setup <- setup_test_values(df_name = "df", col_name = "y", values = c("X", "Y"))
}
