% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bool.R
\name{test_all_false}
\alias{test_all_false}
\title{Test all values equal false}
\usage{
test_all_false(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
vector
}
\description{
Test all values equal false
}
\examples{
vec <- c(T, F, T, F)
df <- data.frame(col1 = c(F, F, F, F))

test_all_false(vec)[2]
## Returns ERROR

test_all_false(df$col1)[2]
## Returns PASS
}
\seealso{
Other boolean tests: \code{\link{test_all_true}},
  \code{\link{test_any_false}}, \code{\link{test_any_true}}
}
\concept{boolean tests}
