% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bool.R
\name{test_all_false_test}
\alias{test_all_false_test}
\title{Test all values equal false}
\usage{
test_all_false_test(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
boolean
}
\description{
Test all values equal false
}
\examples{
vec <- c(TRUE, FALSE, TRUE, FALSE)
df <- data.frame(col1 = c(FALSE, FALSE, FALSE, FALSE))

test_all_false(vec)
## Returns FALSE

test_all_false(df$col1)
## Returns TRUE
}
