% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bool.R
\name{test_all_true}
\alias{test_all_true}
\title{Test all values equal true}
\usage{
test_all_true(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
vector
}
\description{
Test all values equal true
}
\examples{
vec <- c(T, F, T, F)
test_all_true(vec)[2]
# FAILS
df <- data.frame(col1 = c(T, T, T, T))
test_all_true(df$col1)[2]
# PASSES
}
\seealso{
Other boolean tests: \code{\link{test_all_false}},
  \code{\link{test_any_false}}, \code{\link{test_any_true}}
}
\concept{boolean tests}
