% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bool.R
\name{test_all_true_test}
\alias{test_all_true_test}
\title{Tests all values equal true}
\usage{
test_all_true_test(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
boolean
}
\description{
Tests all values equal true
}
\examples{
\dontrun{
vec <- c(TRUE, FALSE, TRUE, FALSE)
df <- data.frame(col1 = c(TRUE, TRUE, TRUE, TRUE))
 
test_all_true_test(vec)
## Returns FALSE

test_all_true_test(df$col1)
## Returns TRUE
}
}
