% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bool.R
\name{test_any_false}
\alias{test_any_false}
\title{Tests any values equal FALSE}
\usage{
test_any_false(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
vector
}
\description{
Tests any values equal FALSE
}
\examples{
vec <- c(T, F, T, F)
df <- data.frame(col1 = c(T, T, T, T))

test_any_false(vec)[2]
## Returns PASS

test_any_false(df$col1)[2]
## Returns ERROR
}
\seealso{
Other boolean tests: \code{\link{test_all_false}},
  \code{\link{test_all_true}}, \code{\link{test_any_true}}
}
\concept{boolean tests}
