% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bool.R
\name{test_any_false_test}
\alias{test_any_false_test}
\title{Tests any values equal FALSE}
\usage{
test_any_false_test(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
boolean
}
\description{
Tests any values equal FALSE
}
\examples{
vec <- c(T, F, T, F)
test_any_false(vec)
# PASSES
df <- data.frame(col1 = c(T, T, T, T))
test_any_false(df$col1)
# FAILS
}
