% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bool.R
\name{test_any_true}
\alias{test_any_true}
\title{Tests any values equal TRUE}
\usage{
test_any_true(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
vector
}
\description{
Tests any values equal TRUE
}
\examples{
vec <- c(T, F, T, F)
df <- data.frame(col1 = c(T, T, T, T))

test_any_true(vec)[2]
# PASSES

test_any_true(df$col1)[2]
# PASSES
}
\seealso{
Other boolean tests: \code{\link{test_all_false}},
  \code{\link{test_all_true}}, \code{\link{test_any_false}}
}
\concept{boolean tests}
