% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_bool.R
\name{test_any_true_test}
\alias{test_any_true_test}
\title{Tests any values equal TRUE}
\usage{
test_any_true_test(vec)
}
\arguments{
\item{vec}{the vector or column to test}
}
\value{
boolean
}
\description{
Tests any values equal TRUE
}
\examples{
\dontrun{
vec <- c(TRUE, FALSE, TRUE, FALSE)
df <- data.frame(col1 = c(TRUE, TRUE, TRUE, TRUE))

test_any_true_test(vec)
## Returns TRUE

test_any_true_test(df$col1)
## Returns TRUE
}
}
