% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_conditional}
\alias{test_conditional}
\title{Test conditional}
\usage{
test_conditional(setup)
}
\arguments{
\item{setup}{the setup class}
}
\description{
Applies the attribute of a passing or failing test
}
