% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_numeric_col.R
\name{test_double_range}
\alias{test_double_range}
\title{Test double range}
\usage{
test_double_range(df, col_name, upper_inclu, lower_inclu, upper, lower, na)
}
\arguments{
\item{df}{the dataframe}
}
\description{
Tests the double range based on the parameter configurations
}
