% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_setup.R
\name{test_exclu_lower}
\alias{test_exclu_lower}
\title{Test Exclu lower}
\usage{
test_exclu_lower(df, setup)
}
\arguments{
\item{df}{the dataframe}

\item{setup}{the setup class for testing}
}
\description{
Tests that column is greater than the lower bound
}
