% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{test_expc_params}
\alias{test_expc_params}
\title{Test the expected parameters exist in the class}
\usage{
test_expc_params(setup, expc_params)
}
\arguments{
\item{setup}{the class to check for the parameters in}

\item{expc_params}{the expected parameters}
}
\description{
Test the expected parameters exist in the class
}
