% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_setup.R
\name{test_inclu_lower}
\alias{test_inclu_lower}
\title{Test inclu lower}
\usage{
test_inclu_lower(df, setup)
}
\arguments{
\item{df}{the dataframe}

\item{setup}{the setup class for testing}
}
\description{
Tests that column is greater than or equal to the lower bound
}
