% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_range.R
\name{test_inclu_lower_exclu_upper}
\alias{test_inclu_lower_exclu_upper}
\title{Test inclu lower and exclu upper}
\usage{
test_inclu_lower_exclu_upper(df, setup)
}
\arguments{
\item{df}{the dataframe}

\item{setup}{the setup class for testing}
}
\description{
Tests that the column is greater than or equal to the lower bound AND less than the upper bound
}
