% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_na.R
\name{test_na}
\alias{test_na}
\title{Test NA values}
\usage{
test_na(df, setup)
}
\arguments{
\item{df}{the dataframe}

\item{setup}{the setup setup for testing}
}
\description{
Tests if the vector/column contains any NA values
}
\examples{
\dontrun{
df <- data.frame(x = 1:4, y = c(NA, 6:8))
setup <- setup_test_na("df_name", "x") 
test <- test_na(df, setup)
## test$test_result returns TRUE

setup <- setup_test_na("df_name", "y") 
test <- test_na(df, setup)
## test$test_result returns FALSE
}
}
