% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{test_param_string}
\alias{test_param_string}
\title{Test param is string}
\usage{
test_param_string(setup, param)
}
\arguments{
\item{setup}{the class to check the parameters}

\item{param}{string, the parameter to check}
}
\description{
Test param is string
}
