% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_col.R
\name{test_unique}
\alias{test_unique}
\title{Test uniqueness}
\usage{
test_unique(df, col_name)
}
\arguments{
\item{df}{the dataframe}

\item{col_name}{the column name}
}
\description{
Tests if the vector/column is unique
}
