% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_unique.R
\name{test_unique}
\alias{test_unique}
\title{Test uniqueness}
\usage{
test_unique(df, setup)
}
\arguments{
\item{df}{the dataframe}

\item{setup}{the setup setup for testing}
}
\description{
Tests if the vector/column is unique
}
\examples{
\dontrun{
df <- data.frame(x = 1:4, y = c(1,1,2:3))
setup <- setup_test_unique("df_name", "x", FALSE) 

test <- test_unique(df, setup)
## test$test_result returns TRUE

setup <- setup_test_unique("df_name", "y", FALSE) 
test <- test_unique(df, setup)
## test$test_result returns FALSE
}
}
