% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_col.R
\name{test_values}
\alias{test_values}
\title{Test expected values}
\usage{
test_values(df, col_name, values, na)
}
\arguments{
\item{df}{the dataframe}

\item{col_name}{the column name}

\item{values}{the expected values for the variable}

\item{na}{if the column should include NA values or not, takes the values TRUE or FALSE}
}
\value{
class with test_result, test_message, test_description
}
\description{
Tests if the vector/column contains values other than expected
}
\examples{
\dontrun{
df <- data.frame(x = 1:4, y = 5:8)
values <- 1:4
na <- FALSE

test <- test_values(df, "x", values, na)
## test$test_result returns TRUE

test <- test_values(df, "y", values, na)
## test$test_result returns FALSE
}
}
