% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_col.R
\name{test_values}
\alias{test_values}
\title{Test expected values}
\usage{
test_values(vec, expc_values)
}
\arguments{
\item{vec}{the vector or column to test}

\item{expc_values}{a vector of expected values}
}
\value{
vector
}
\description{
Tests if the vector/column contains values other than expected
}
\examples{
vec1 <- c(1,2,3)
vec2 <- c(1,1,2)
expc_values <- c(1, 2)

test_values(vec2, expc_values)[2]
## Returns PASS

test_values(vec1, expc_values)[2]
## Returns ERROR
}
\seealso{
Other general column tests: \code{\link{test_null_values}},
  \code{\link{test_unique}}
}
\concept{general column tests}
