% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_values.R
\name{test_values}
\alias{test_values}
\title{Test expected values}
\usage{
test_values(df, setup)
}
\arguments{
\item{df}{the dataframe}

\item{setup}{the setup setup for testing}
}
\value{
setup with test_result, test_message, test_description
}
\description{
Tests if the vector/column contains values other than expected
}
\examples{
\dontrun{
df <- data.frame(x = 1:4, y = 5:8)
values <- 1:4
na <- FALSE
setup <- setup_test_values("df_name", "x", values, na)

test <- test_values(df, setup)
## test$test_result returns TRUE

setup <- setup_test_values("df_name", "y", values, na)
test <- test_values(df, setup)
## test$test_result returns FALSE
}
}
